+++ eResourceWeb/Models/ResourceTypeMasterModel.cs
@@ -12,9 +12,9 @@ namespace eResourceWeb.Models
     {
 
         [Key]
-        public Double Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
-        public Double IsHuman { get; set; }
+        public int IsHuman { get; set; }
 
     }
 }
